feat: rename Citadel to Truv
