+++ app/src/androidTest/java/com/example/truvpoc/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.citadelpoc
+package com.example.truvpoc
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.citadelpoc", appContext.packageName)
+        assertEquals("com.example.truvpoc", appContext.packageName)
     }
 }
