+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.citadelpoc">
+    package="com.example.truvpoc">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -12,9 +12,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.NoActionBar"
         android:usesCleartextTraffic="true">
-        <activity android:name=".DisplayEmploymentActivity"></activity>
-        <activity android:name=".DisplayIncomeActivity"></activity>
-        <activity android:name=".MainActivity">
+        <activity android:name="com.example.truvpoc.DisplayEmploymentActivity"></activity>
+        <activity android:name="com.example.truvpoc.DisplayIncomeActivity"></activity>
+        <activity android:name="com.example.truvpoc.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
