+++ app/src/main/assets/www/index.html
@@ -1,22 +1,22 @@
 <html>
     <head>
-        <script src="https://cdn.citadelid.com/bridge.js"></script>
+        <script src="https://cdn.truv.com/bridge.js"></script>
     </head>
     <body>
     <script>
       (async function () {
         const params = new URLSearchParams(window.location.search)
         const bridgeToken = params.get("bridge_token")
-        const bridge = CitadelBridge.init({
-          clientName: 'Citadel Quickstart',
+        const bridge = TruvBridge.init({
+          clientName: 'Truv Quickstart',
           bridgeToken: bridgeToken,
           product: 'employment',
           origin: '*',
           trackingInfo: 'any data for tracking current user',
-          onLoad: citadelInterface.onLoad,
-          onEvent: citadelInterface.onEvent,
-          onSuccess: citadelInterface.onSuccess,
-          onClose: citadelInterface.onClose,
+          onLoad: truvInterface.onLoad,
+          onEvent: truvInterface.onEvent,
+          onSuccess: truvInterface.onSuccess,
+          onClose: truvInterface.onClose,
         });
         window.bridge = bridge;
       })();
