+++ app/src/main/java/com/example/truvpoc/DisplayIncomeActivity.kt
@@ -1,16 +1,15 @@
-package com.example.citadelpoc
+package com.example.truvpoc
 
 import android.os.Bundle
-import android.util.Log
 import android.widget.EditText
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import org.json.JSONObject
 
-class DisplayEmploymentActivity : AppCompatActivity() {
+class DisplayIncomeActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_display_employment)
+        setContentView(R.layout.activity_display_income)
 
         if (intent.hasExtra("verification")) {
             val verification = JSONObject(intent.getStringExtra("verification"))
@@ -37,6 +36,7 @@ class DisplayEmploymentActivity : AppCompatActivity() {
         setEditValue(R.id.textEndDate, employment.getString("end_date"))
         setEditValue(R.id.textPositionType, employment.getString("job_type"))
         setEditValue(R.id.textTitle, employment.getString("job_title"))
+        setEditValue(R.id.textAnnualIncome, employment.getString("annual_salary"))
     }
 
     fun setEditValue(element: Int, value: String) {
